refactor(backend): 이동한 유틸 파일명을 모듈 네이밍 컨벤션에 맞게 수정
- wsRawData.ts → ws-raw-data.util.ts
- directoryParser.ts → directory-parser.ts
- directoryParser.test.ts → directory-parser.test.ts
