+++ backend/modules/files/directory-parser.ts
